Break up the picker
